services/security/requirements.txt
flask==2.3.3
gunicorn==20.1.0
requests==2.31.0
cryptography==41.0.3
pyjwt==2.8.0
bcrypt==4.0.1
python-jose==3.3.0
itsdangerous==2.1.2